+++ package.json
@@ -58,6 +58,8 @@
   },
   "dependencies": {
     "bootstrap": "^3.3.4",
+    "jquery": "^2.1.3",
+    "lodash": "^3.6.0",
     "newforms": "^0.12.1",
     "react": "^0.13.1",
     "react-bootstrap": "^0.20.0",
