+++ src/lib/config.js
@@ -1,0 +1,8 @@
+"use strict";
+
+let Config = {
+  API_URL: "http://localhost:3000/"
+};
+
+window.config = Config;
+export default Config;
